refactor: modularize render analysis and size checks
