+++ tools/analysis/code/analyze-render-performance/functions/count-matches.ts
@@ -1,0 +1,3 @@
+export function countMatches(content: string, regex: RegExp): number {
+  return content.match(regex)?.length ?? 0
+}
