+++ tools/misc/metrics/enforce-size-limits/functions/count-lines-of-code.ts
@@ -1,0 +1,9 @@
+export function countLinesOfCode(content: string): number {
+  return content
+    .split('\n')
+    .filter(line => {
+      const trimmed = line.trim()
+      return trimmed.length > 0 && !trimmed.startsWith('//')
+    })
+    .length
+}
