+++ tools/misc/metrics/enforce-size-limits/functions/max-nesting-depth.ts
@@ -1,0 +1,15 @@
+export function maxNestingDepth(content: string): number {
+  let maxDepth = 0
+  let currentDepth = 0
+
+  for (const char of content) {
+    if (char === '{' || char === '[' || char === '(') {
+      currentDepth++
+      maxDepth = Math.max(maxDepth, currentDepth)
+    } else if (char === '}' || char === ']' || char === ')') {
+      currentDepth--
+    }
+  }
+
+  return maxDepth
+}
